Add index property to gcc.BasicBlock, and show it in the CFG graphviz output
The repr() of a gcc.GimplePhi (from gimple-pretty-print.c:dump_gimple_phi)
contains all of the possible values, and each such arg's representation
contains the index of the source block for the relevant incoming edge, in
parentheses.

For example:
   i_1 = PHI <i_4(2), i_11(3)>
shows that the args correspond to edges coming from blocks 2 and blocks 3.

Display the index of each block in the CFG visualization to make it easier to
read the PHI representations.