fix: permerror() misusage
The GCC permerror() function takes at least two arguments, with the
second one being a formatted litteral. The current usage in gcc-python.c
of this function is to pass directly as a second argument a string and
not a litteral. The present commit fixes the issue by passing a litteral
"%s" and using the msgid as a formatting parameter. This also fixes
compilation when using -Werror=format-security (default of Mandriva).