Work-in-progress on refcount checker
Do the checking earlier, on the pre-SSA representation, rather than on the
SSA, so that we have more user-readable variable names (this means we're
dealing with gcc.VarDecl rather than gcc.SsaName)

Use OrderedDict within states to preserve a meaningful ordering of the
components.

Introduce a Transition class to support associated descriptive text with a
state transition; reword accordingly

Rather than a depth-first traversal of an implicit graph of states, instead
we generate an explicit graph (the StateGraph and StateEdge classes).

Add visualizations of these state graphs to help with debugging the checker
(libcpychecker.visualizations)