Expose more of the pass machinery, and use this to include the full pass tree within the docs
Expose the pass PROP_* bitfield flags as constants within the gcc module.

Add "next" and "sub" properties to gcc.Pass

Add gcc.Pass.get_roots() to get at the top-level passes.

Add documentation for the above.

Add generate-tables-of-passes-rst.py, which uses the above to generate tables
of passes for the documentation.

Note that docs/tables-of-passes.rst is autogenerated, but we track it in GIT
to avoid needing to build the plugin to build the docs.