Detect reference leaks for values that aren't the return value
Introduce a ConcreteValue subclass of AbstractValue, and eliminate the
NullPtrValue class, so that we can treat "NULL" in C code in the same way
that we deal with simulated function-call failures

Track the location of AbstractValue, rather than the stmt, in the hope that
this will be available in more places.

Rewrite the ob_refcnt tracker so that it investigates all ob_refcnt fields
we know something about at the end of each trace, rather than just that of the
return value.

Add a new test case:
  tests/cpychecker/refcounts/object_leak
to verify detection of leaking an object reference within a function.