Avoid hardware dependent results in the tests/plugin/parameters selftest
Out of all the parameters I could have chosen for the selftest, I managed
to pick the two parameters that vary based on the hardware on the test
machine, thus embedding an expected value that only worked with a specific
certain amount of RAM.

Specifically, 'ggc-min-expand' and 'ggc-min-heapsize' are set up by gcc's
init_ggc_heuristics() to values based on the amount of physical memory on the
test machine.

Change it to something that (hopefully) has a hardware-independent default.