Generate HTML reports on errors (in addition to text on stderr)
Add new file:
  libcpychecker/diagnostics.py
containing a higher-level interface on top of gcc.error() et al for capturing
information about an error.

Add HtmlRenderer class, which can dump the captured information in HTML/CSS/JS
form, using pygments to syntax highlight the code, and the jsplumb JavaScript
library to add control-flow arcs (via the HTML <canvas> element) to the
result.

For now, any refcount errors in each function are all gathered and dumped into
a single file (per function), named:
  name-of-function-refcount-errors.html

The relevant JavaScript libraries are referenced by URL within the generated
HTML, to keep things selfcontained, and to avoid dealing with packaging them
with the plugin.

The resulting reports work for me with Chromium 12 and Firefox 4

This is all in addition to the regular textual messages on stderr