Fix segfault seen on i686 due to erroneous implementation of 'pointer' attribute of gcc.TypeDecl
gcc_TypeDecl_get_pointer() was calling:

   gcc_python_make_wrapper_tree(build_pointer_type(self->t));

where (tree)t had TREE_CODE() == TYPE_DECL, and thus tree_code_type[TYPE_DECL]
is tcc_declaration.

build_pointer_type() assumes the tree passed in is tcc_type

Somehow this was surviving on x86_64, but on i686 it segfaulted during the
lookup of the "may_alias" attribute (on the type_decl-being-treated-as-a-type).

On x86_64, gcc.TypeDecl.pointer was returning a pointer to a thing that was
being treated as a type.

With this change, gcc.TypeDecl.pointer now returns a type that is a pointer to
an actual type.  This changes the output of various selftests, in that we're
printing the string representation of a pointer to the underlying type, rather
than that of a pointer to a typedef (and having this happen to work on some
builds of gcc, despite the violation of gcc's internal type system).

Also, fix the signature of a couple of attribute getters:
  gcc_FunctionType_get_argument_types
  gcc_IntegerConstant_get_constant