Get rid of the PyRuntime class
PyRuntime was an attempt to run the tests against multiple pythons.

However, it was clumsy, and contained hardcoded values (for python 2.7)

This patch removes it, instead simply building with sys.executable

Doing so means that we're now building and running the selftests against
the version of python that the plugin was built against, rather than
hardcoding /usr/bin/python2.7

In particular, when building the plugin for Python 3, it means that the
selftests are now testing compiling against the Python 3 headers.

Doing this located some real bugs for this case, which this patch also
fixes.