cpychecker: fix the handling of "return;" in void-returning functions
Previously we were using "state.return_rvalue is None" to signify that a
function had not yet returned, but this doesn't cover the case of a plain

  return;

statement.

Convert state.has_returned from a method to boolean data, to track this
separately.

(I also tried just setting the new state's location's bb to be the cfg.exit,
but this interferes with error reports, when reporting the location of where
the return statement occurred)