libcpychecker: Improvements to handling of globals and of pointers
Handle references to parameters, as well as locals

The first time we look up the value of a global, assign it a new "unknown"
value.

Generalize NullPtrDereference into cases which are definite vs those that are
merely possible

Introduce "SplitValue" exception, using it to support backtracking the
evaluation of a State, splitting an UnknownValue within the state into multiple
alternate AbstractValue instances that we know more about.

Split comparison of global pointers with NULL into parallel worlds in which the
pointer is non-NULL and those in which it's NULL.

Split lookup of a field through a pointer into parallel worlds for the NULL vs
non-NULL case

Rework loop detection so that it handles split states.

New test cases, covering a read through a global ptr (that might be NULL), and
a read through a global ptr after checking that it's non-NULL