cpychecker: revamp of reference-count checking
Split up the counting within the RefcountValue class to track both those
references that the function "owns" versus those owned by the rest of the
program state.  This allows us to know for an object e.g. that the function
doesn't directly own any references, but that the reference count is still > 0
(for a borrowed reference), as compared to the case where the function owns a
reference, but we don't know of any in the rest of the program (typical for a
"new reference" e.g. from a constructor).

Explicitly track which regions are on the stack (by being directly marked as
such, or by being a child of a stack region)

Within impl_PyList_New, create an "ob_item" array in the simulated heap, in
case the code modifies it.

Within the reference-count checker, we now look for memory locations that
store references to objects.   If those locations not on the stack, then the
references they store are now assumed to legally count towards the ob_refcnt
that the function "owns".  This is needed in order to correctly handle the
PyList_SET_ITEM() macro, which directly writes to the list's ob_item field,
"stealing" a reference: we can detect these references, and count them towards
the ob_refcnt value.

Extend RefcountAnnotator so that it reports on "persistent" reference changes
to the object in question (i.e. within non-stack memory locations)

When reporting on a reference-counting error, report on what we thought the
reference count was, and what we thought it should have been, listing all of
the references we know about (e.g. return value, stored pointers on the heap
or within globals).

Add test cases that use the PyList_SET_ITEM macro, in order to exercise the
above, and update the expected test outputs to reflect the above changes.

[all of the above was motivated by needing to correctly handle
the PyList_SET_ITEM() macro]