cpychecker: improve handling of conditionals; add testcase support functions
Special-case handling for calls to certain functions to help with debugging
and selftests:

    __cpychecker_log()

       Log the given string constant within the trace

   __cpychecker_dump()

       Log the given r-value within the trace

   __cpychecker_assert_equal()

      Compare two r-values and raise an AssertionError if they are non-equal

Implement non-equality checking for ConcreteValue to help with the above
assertion checks

Refactor State.eval_condition() so that it's usable both for GimpleCond
statements and for eval_rhs (boolean expressions).

Extend it from supporting just == and != to add <, <=, >, >=

Add test cases for the conditional support (using the new __cpychecker
functions).