cpychecker: refinements to handling of NULL pointers
Previously, any time we dereferenced an unknown pointer value, we considered
the path in which it was NULL, and warned about a possible NULL pointer
dereference.

This leads to large numbers of false positives when running on real-world
code, so for now the default is not to report on those errors, and assume that
such code is correct: that the unknown pointer value was not NULL.

This is a boolean, so that it can potentially be a command-line option within
the invocation harness.

Note that for the cases where we know that the pointer can _definitely_ be NULL,
an error will still be reported (e.g. when considering the various possible
return values for a function known to be able to return NULL).