cpychecker: fix set_exception method
MyState.set_exception was erroneously setting self.exception_rvalue to the
Region of the named exception, rather than the value of a pointer to it.

Add a "loc" argument so that the location where the exception is set can be
propagated through to the PointerToRegion value, and update all calls.