cpychecker: improve the descriptions of state transitions
The various Transition.desc values are no longer printed with a "when " prefix.

Instead, we determine on a case-by-case basis whether or not there are
multiple possibilities, and only use "when " if there's more than one choice.

For example, a function call might have two transitions with these descriptions:

   "when PyLong_FromLong() succeeds"
   "when PyLong_FromLong() fails"

whereas a conditional for which it's provable that only one path is possible
might have this description:
   "taking True path"