cpychecker: implement gcc.ConvertExpr for WithinRange instances
Implement type conversion for WithinRange.

Initially, I attempted to emit errors for the case where a value could be
truncated (e.g. from long to char).

However this appears not to be possible to do without generating lots of
noise.  For example, gcc will convert:

   ch = i & 0xff  (char ch, int i)

to just:

   ch = (char)i

before our python code sees it.

Given that it doesn't seem possible for us to distinguish between an explicit
and an implicit conversion, we instead simply track the case where truncation
could happen by making it be any value in the new type, and pass on data that
is known to be within the range of the new type.