cpychecker: reworking of l-values and r-values
Extracting a field from a struct with unknown value revealed a flaw in the
existing scheme for handling r-values: the UnknownValue() would be inherited
from the parent region (the struct), but would have the wrong type:
  UnknownValue(gcctype=some_struct)
rather than
  UnknownStruct(gcctype=type_of_field)

This lead to various flaws, including Python exceptions when handling pointer
fields, since the machinery for splitting the r-value into the NULL/non-NULL
cases requires such an r-value to be a pointer type.

This is covered by new test cases:
  tests/cpychecker/absinterp/nested-fields2
  tests/cpychecker/absinterp/nested-fields3

This patch introduces an AbstractValue.extract_from_parent() method, which can
be called on a r-value supplied as a default by a parent region, coercing it to
the appropriate type for the child region.

Doing so uncovered an issue with array handling.  Within:
   int arr[10];
   int *ptr = arr;

ptr's r-value had type "&int[10]", rather than type "int*".  This led to
issues where e.g.:

   ptr[0] = 42;
   return ptr[1];

would leak to 'arr' being assigned a default r-value of 42, when really only
'arr[0]' should have the r-value, and thus ptr[1] would erroneously evaluate
to 42, rather than be detected as uninitialized.

The fix for this is to introduce a new ArrayElementRegion subclass, signifying
a specific indexed location within an array, and for the above to give ptr the
r-value "&arr[0]" (with type 'int*') rather than "arr", and to update pointer
arithmetic accordingly.

These test cases add additional test coverage for pointer arithmetic within an
array:
  tests/cpychecker/absinterp/arrays5
  tests/cpychecker/absinterp/arrays6

The introduction of ArrayElementRegion has a knock-on effect on how we handle
string constants: in code of this form:

   char *str = "hello world";

'str' now with an r-value of &("hello world"[0]) since this is a char*, rather
than of the "hello world" directly, since this is a char[12]. (i.e. it has a
PointerToRegion() to a ArrayElementRegion() within aRegionForStringConstant(),
rather than to the latter directly).