cpychecker: fix false warning about uninitialized data for arrays on the stack
cpychecker was issuing false warnings for some cases where an array is on the
stack and is either fully or partially initialized, when the index is not
fully known.

This was happening due to the array lookup giving us an ArrayElementRegion
within a unknown index (a WithinRange instance).  This led to the vaue lookup
deferring to the array's default r-value, which is an UninitializedValue().

We now detect this case, and instead determine what we can say (if anything)
about the possible elements that we could be looking up.