cpychecker: allow the checker to run on an SSA representation
This patch adds minimal support to libcpychecker to allow it to run on the
SSA representation of a function.

However, it doesn't change when the "cpychecker" pass runs (actually changing
this causes numerous changes to the output of the self tests, alas).

Note that we don't really do much with the SSA form: we ignore Phi nodes, and
all gcc.SsaName() references are handled by referring back to the underlying
variable, so that our region-to-value and pointer-to-region mappings should
continue to work.

However, in theory this may allow us e.g. to use facts about the code deduced
by SSA optimization passes.