cpychecker: add a custom "cpychecker_returns_borrowed_ref" attribute
Add a custom:
  cpychecker_returns_borrowed_ref
attribute to the checker, along with the pre-defined preprocessor name:
  WITH_CPYCHECKER_RETURNS_BORROWED_REF_ATTRIBUTE
so that Python C code can mark functions returning a PyObject* for which the
returned reference is borrowed, rather than a new reference.

Update the checker to take this into account when considering both calls to
those function, and when analyzing such functions themselves.