cpychecker: fix handling of loops
Previously, we stopped analyzing a trace if we reached a Location that we
had been to before, with some special-casing for split states, and for calls
to functions that don't return, like exit(), with the aim of analyzing zero
and one times through each loop.

This turned out not to be working as desired: it was silently discarding all
traces containing a loop.  For example, given

   /* (A): before loop */

   for (i = 0; i < n; i++) {

      /* (B): body of loop */

      if (some_condition) {
          goto error;
      }
   }

   /* (C): after loop */

 error:
   /* (D): cleanup */

the checker would cover these traces:

  (A) -> (C)  (zero iterations)
  (A) -> (B) -> (D)   (one iteration, interrupted by the goto)

but the trace  (A) -> (B) -> (C) would be discarded, since the checker
would reach the   i < n   conditional for a second time, and throw away
that trace.

This patch changes the loop detection to instead work on transitions, similar
to edges in the CFG, but covering pairs of (src, dest) Location instances.

We now detect a loop and terminate the analysis for a trace that has followed
a (src, dest) path for a second time.  In the above example, this means that
it's now OK to execute the (i<n) conditional a second time, but the checker
will stop for the trace where we go from that conditional into the loop body a
second time.  This means that the checker _will_ now cover the (A) -> (B) -> (C)
trace (i.e. once through the loop).

So in theory the checker ought to now analyze the Trace instances where a
loop body is executed zero and one times.