cpychecker: de-duplicate error reports
Implement simple de-duplication of error reports.

Add a remove_duplicates() method to the Reporter class, to walk the gathered
Report instances, and partition them into equivalence classes before flushing
them.  Only the first Report within each such partition is flushed (together
with a note about the number of duplicates seen).

For now we implement only a simple equivalence relationship: two reports are
equivalent if they are for the same function, at the same location, and with
the same summary text.

This successfully removes a genuine duplicate within the test suite:
  tests/cpychecker/refcounts/PyDict_SetItem/incorrect
and when compiling the plugin using gcc-with-cpychecker