cpychecker: Py_BuildValue("O") can accept a ptr to any PyObject subclass
The checker for Py_BuildValue("O") (and "S" and "N") was being too strict,
requiring a (PyObject*).  Although it's not explicitly documented, it's
clear that these can also accept pointers to any PyObject subclass.

Fixes a false positive seen when running cpychecker on coverage-3.5.1b1.tar.gz,
in which coverage/tracer.c:Tracer_trace passes a PyFrameObject* as an argument
to such a call.