cpychecker: start adding explanations to NULL function argument warnings
Add an optional "why" string attribute to NullPtrArgument, setting
these up as appropriate for many of the places where we check this.

Doing so means that we can directly tell the user why we're complaining
about a NULL being passed to a function, rather than hiding the
rationale within a comment in the cpychecker source code, leading to
more readable error reports (I hope).