cpychecker: use fnmeta.name
Now that we have FnMeta instance, use the 'name' attribute, rather than
duplicating this information, or by accessing it from stmt.fn.operand.name

This helps to reduce the number of places where we look up a function name
via 'stmt.fn.operand.name'.  The latter works for regular function callsites,
but makes the assumption that the gcc.GimpleCall has a 'fn' that's a
gcc.AddrExpr, which isn't the case when jumping through a function pointer,
since we'd have a gcc.VarDecl for the 'fn' instead.