cpychecker: implement PyList_Size and improve field lookups
Implement State.read_field_by_name() to make it easy to get fields from a
structure in a sane way that's consistent even if we're only working with
unknown values.

Use this to implement PyList_Size

Update UnknownValue.is_equal() and WithinRange.is_equal() so that a specific
UnknownValue is equal to itself.

tests/cpychecker/refcounts/PyList_Size/known-size:
  This selftest verifies that we can look up a field we set earlier

tests/cpychecker/refcounts/PyList_Size/unknown-size:
  This selftest verifies that if we read an unknown field twice, we get the
  *same* value.

This ensures that if we call PyList_Size() twice on a list, we get consistent,
equal values (unless something changes the size, of course).