cpychecker: improve behavior on functions that are too complicated to fully handle
Previously (as of d667b9ed430f9df5f4c65fa977a4a30091573101), when a function
was too complicated for the reference-count tracker to fully analyze (based
on the number of state transitions seen), it would give up, performing no
analysis.

With this commit, we instead try to obtain at least some subset of the list
of all traces through the function, and analyze those.  Given that we do a
depth-first traversal of the tree, and that "success" transitions are typically
visited before "failure" transitions, this means that we should at least see
the trace in which all functions calls succeed, together with traces in which
some of the later calls fail.

Add a selftest:
  tests/cpychecker/refcounts/combinatorial-explosion-with-error
which has a leak in the "success" case, to verify that we report on this
error.

Unfortunately, the semantics of PyModule_AddObject are poorly specified.
According to:
  http://docs.python.org/c-api/module.html#PyModule_AddObject
"This steals a reference to value. Return -1 on error, 0 on success." but
in fact a reference is only stolen on success.

This means that the new selftest also has numerous refcounting errors for
the case where one or more of the PyModule_AddObject calls fails (it leaks
one reference per failed called), as seen in the expected output, but these
problems reflect those seen in real-world code.

The commit rewrites the existing selftest's code:
  tests/cpychecker/refcounts/combinatorial-explosion/input.c
to properly handle failures within the PyModule_AddObject calls, so that
there are no expected reference leaks within this code.