cpychecker: add bulletproofing to refcount modifications
In many cases, PyObject* values are initialized by the checker to have
an ob_refcnt, but it's possible for CPython.change_refcount() to
encounter a PyObject* for which the ob_refcnt is unknown.

This patch makes it treat such a case as a borrowed reference, rather
than failing with a traceback.

Fixes a traceback seen when running the checker on rpm's python bindings
in python/header-py.c:hdr_reduce where we don't know the ob_refcnt of
Py_TYPE(s) in:
    Py_BuildValue("O(O)", Py_TYPE(s), blob);