support emitting unconditional warnings
Make the "option" argument to gcc.warning be optional: if not provided, an
unconditional warning is emitted.

I believe this argument will be rarely used (until we can add new command-line
options), so it makes sense to make it optional, moving it to the end of the
positional arguments.

This reorders the arguments to gcc.warning, but there are no uses of this
entrypoint within the source tree (other than test code), and a review of all
code posted by others to the mailing list showed no usage of the entrypoint
either.

Fixes ticket #8:

  https://fedorahosted.org/gcc-python-plugin/ticket/8 */

Also, fix the order of the command-line arguments to gcc-with-python in the
usage examples: some erroneously contained gcc options before the python
script name (which must come first).