cpychecker: protect against division-by-zero python tracebacks (ticket #25)
Protect against Python tracebacks with division by zero

   ZeroDivisionError: integer division or modulo by zero

by catching such python exceptions, and converting to a new PredictedError
subclass that wraps the underlying arithemetic error.

Also, when encountering code like this:

    int
    test(int i, unsigned int j)
    {
        return i % j;
    }

where j is WithinRange(0, UINT_MAX), split such a WithinRange into the 0 and
the >0 cases first, so that the division by zero only happens in the former
case.

Fixes the Python traceback seen in:
   https://fedorahosted.org/gcc-python-plugin/ticket/25
though there is still a warning, due to considering the (j == 0) case.