cpychecker: fix python traceback analyzing PyList_Append(list, UnknownValue() ) (ticket #22)
Python tracebacks were seen when analysing code within gdb of the form:
   PyList_Append(list, item)
where item is an UnknownValue.

The root cause is the family of methods of CPython which call
change_refcount(pyobjectptr), which requires a PointerToRegion.

This patch updates it to handle an UnknownValue by first splitting the state
into the NULL and non-NULL worlds, so that next time the latter is a
PointerToRegion (and which eeb4d89f6686864e84f085a971f328a44524f305 will treat
as being a borrowed reference).

Fixes https://fedorahosted.org/gcc-python-plugin/ticket/22