cpychecker: fix python traceback analyzing PyList_Append() (ticket #19)
Python tracebacks were seen when analysing code within gdb of the form:
  PyList_Append(list, item)
where list is an UnknownValue.

The initial problem was that the check for a NULL argument was only
checking against a known NULL value (ConcreteValue).  This patch updates
State.raise_any_null_ptr_func_arg() so that if it receives an UnknownValue,
it splits the world into the NULL and non-NULL versions.  Doing so adds
various new traces to existing test cases, covering the worlds in which those
arguments are NULL, which would add new errors, but we also update the checker
so that it doesn't report NULL argument errors that are from such a state
split, since they're not proven (analogous to
5b98cec10143271d6365951f5d3037f04c123159).

Fixing this led to another problem: such an unknown PyListObject doesn't have
a known ob_size, which makes it difficult to append a reference at some
specific place within the ob_item array.  We get around this by detecting
this case, and making an arbitrary assumption that it's the smallest
non-negative size currently possible (e.g. 0).  To do so, we reuse the
SplitValue machinery so that we can make a note of the assumption in the
trace's description text, adding a new WithinRange.raise_as_concrete() method.

Fixes https://fedorahosted.org/gcc-python-plugin/ticket/19