cpychecker: fix python traceback with negative shifts (ticket #14)
When encountering a shift by a negative value, split the state into the
(RHS < 0) and (RHS >= 0) worlds, leading to the former being merely a
"possible" shift-by-negative error; don't report arithmetic errors that
are only possible.  Doing so silences the warning within:
  tests/cpychecker/absinterp/arithmetic/division-by-zero/possible/

Fixing this leads to massive overflow of the upper bound, leading to huge CPU
usage when printing the hexadecimal repr of (2**31-1) << (2**31-1), so
clamp WithinRange values to actually be within the expressible range of the
underlying type.  This leads to some changes to existing test cases, where
a Py_INCREF of an unknown refcount was overflowing the Py_ssize_t by 1.

Fixes https://fedorahosted.org/gcc-python-plugin/ticket/14