hide references to gcc.VarDecl in output (ticket #12)
gcc.Declaration was using its __repr__ as its __str__, so that str(decl)
might show e.g. "gcc.VarDecl('foo')"

Remove this, so that it falls back on using the standard gcc.Tree.__str__,
which is implemented in terms of gcc's pretty-printing routines.

This changes various "%s" within libcpychecker so that they no longer print
"gcc.VarDecl()" on variables, instead printing the name of the variables.

Take this opportunity to improve the wording of some of these error messages
(e.g. using get_symbol() for exprcode variables)

Update various other selftests to reflect the change in gcc.Declaration.__str__

This fixes many of the problems seen in
https://fedorahosted.org/gcc-python-plugin/ticket/12

Note that for some selftests I still see warnings containing names that
don't reflect the user's view of the code, e.g.:

  tests/cpychecker/absinterp/nested-fields3
     "dereferencing uninitialized pointer (D.12022->field)"

  tests/cpychecker/absinterp/uninitialized-data
     "comparison against uninitialized data (j.0)"

  tests/cpychecker/refcounts/PyArg_ParseTuple/correct_O_bang:
     "as argument 1 (D.12030)"

  tests/cpychecker/refcounts/PyArg_UnpackTuple/missing-initialization:
     "uninitialized data (b.1)"

  tests/cpychecker/refcounts/PyList_GetItem/correct:
     "argument 1 (some_list.0)"