cpychecker: rewrite of how values are compared
I investigated why the cpychecker was not reporting a particular refcounting
bug seen by eye, and discovered a major bug in how the cpychecker compared
values:

    int i = 0;
    if (i < some_unsigned_value_of_unknown_value) {
        /* this was never reached by the checker */
    } else {
        /* the checker was assuming that this suite was *always* executed */
    }

i.e. it was assuming that
  (int)0 >= any value for an unknown (unsigned int)

i.e.:

   ConcreteValue(int, loc, 0) >= WithinRange(unsigned int, 0, UINT_MAX)

due to boundary errors in how ConcreteValue and WithinRange values were
compared: it was erroneously treating this as being always true, when it is
only true when the RHS has value 0.

To try to shake out such bugs, this patch adds test coverage for the following
comparisons between a ConcreteValue and the boundaries of a WithinRange:
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-eq-max-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-eq-min-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-ge-max-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-ge-min-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-gt-max-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-gt-min-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-le-max-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-le-min-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-lt-max-of-WithinRange
   tests/cpychecker/absinterp/comparisons/boundaries/ConcreteValue-lt-min-of-WithinRange
      (^^^ this is the case described in the pseudocode above)

Doing so revealed that at least 3 of the above comparisons had the wrong
behavior.

Attempting to fix the existing implementation uncovered more bugs, so in the
end I rewrote it.

In the old implementation, AbstractValue had various comparison methods
(is_equal, is_lt, impl_is_lt, impl_is_ge, is_le) implemented in terms of each
other, some of which were implemented by subclasses.

This patch changes the implementation so that there is a single comparison
method to be implemented by each AbstractValue class (eval_comparison), which
takes an operator name as an additional argument ('eq', 'ge', 'gt', 'le',
'lt')

This patch also adds a new WithinRange.make class method and uses it
throughout, so that if a range contains only a single value, we get a
ConcreteValue instead.

The comparison ConcreteValue == WithinRange is now implemented by splitting
the WithinRange against the ConcreteValue giving the too-low-subrange (False),
the value (True) and the too-high-subrange (False), or only 1 or 2 subranges
where we're on the boundary.  Given that the old behavior was to return "None"
for unknown, this changes the behavior of a comparison of a known value against
an unknown value from generating two traces (equal vs non-equal) to three
traces (too low, equal, too high) and adds a textual description of the ranges
to the Transition descriptions.  This changes the number, ordering and
descriptions of the traces in many of the selftests.  Note that both old and
new implementations return False when equality is impossible.

The other comparisons for ConcreteValue vs WithinRange (inequalities) are now
implemented by generating a list of boundary values, running the comparison
against all interesting boundary values, and then splitting the WithinRange
into the minimum number of subranges for which the comparison gives the same
answer across each subrange.

This patch also adds some additional test coverage:
    tests/cpychecker/absinterp/comparisons/boundaries/WithinRange-eq-ConcreteValue-in-middle

There is also a "debug_comparisons" flag within libcpychecker/absinterp.py
which can be toggled on when debugging all of the above.