cpychecker: extend detection of PyObject* subclasses
The checker was treating any struct beginning with "ob_refcnt" and "ob_type"
as a Python object (with some tweaks for python 3 and debug builds).

Extend this to also cover structs that begin with a field that's a PyObject
(or subclass), since these are likely to also be PyObject subclasses.