add metaclass to PyGccWrapper to fix garbage-collection segfault on user-defined classes
Previously, all PyGccWrapper instances assumed that their ob_type was a
PyTypeObject that was also a PyGccWrapperTypeObject, a PyTypeObject extended
to hold a wrtp_mark callback.

This worked for predefined types, where the type object was statically
allocated, but led to crashes with user-defined types, where the new type is
a PyHeapTypeObject: I believe that upon casting to
   *(PyGccWrapperTypeObject*)typeobj)->wrtp_mark
we were actually accessing the as_number field:

   /* The *real* layout of a type object when allocated on the heap */
   typedef struct _heaptypeobject {
       PyTypeObject ht_type;
       PyNumberMethods as_number; <--- this field i.e. the:
       [...snip...]
   } PyHeapTypeObject;

thus receiving the first field of PyNumberMethods:

               binaryfunc nb_add;

(casting it with the wrong signature) which happened to be NULL.

With this change, we increase PyGccWrapperTypeObject to be a PyHeapTypeObject
(even for those that aren't on the heap) so that user-defined types have enough
room.

We introduce a metaclass: PyGccWrapperMetaType, the ob_type of our various
PyGccWrapperTypeObject instances. This has a tp_basicsize that ensures that
subclasses have room for the extra field in PyGccWrapperTypeObject, and a
tp_new implementation that propagates the wrtp_mark callback when creating
new subtypes.

Fixes the issue reported as:
  https://fedorahosted.org/pipermail/gcc-python-plugin/2012-January/000152.html
and adds a reproducer for it as:
  tests/plugin/gc/segfault-on-instance-of-pass-subclass/