add a configuration phase and use it to support PLUGIN_FINISH_DECL
PLUGIN_FINISH_DECL was added in GCC 4.7, and there aren't any ifdefs we can
use to directly test for it (other than version numbers).

Introduce generate-config-h.py to the build, to perform feature testing,
writing out a new autogenerated-config.h file.  (also, introduce a support
file, configbuilder.py, to do most of the work).

Use this to add support for wiring up a callback to PLUGIN_FINISH_DECL, given
qa version of GCC where this hook exists.

Add [WhenToRun] section to metadata.ini in run-test-suite.py, so that tests
can be conditional on particular features being supported, and use this to
add a selftest for PLUGIN_FINISH_DECL that's marked as requiring the feature.