cpychecker: fix an off-by-one in the handling of conditionals
When splitting a WithinRange against a ConcreteValue due to a conditional,
ensure that all of the boundaries are within the existing WithinRange,
and extend the final boundary up to the final range, rather than assume that
the final range covers [boundary[n-1]..boundary[n]] (which isn't true when
the split is happening close to the max value of the range).

Fixes a traceback seen with a pair of loops when compiling
gstreamer-python-0.10.19 under the checker.

Also, make it easier to debug comparisons by by indenting the optional debug
text that appears when debug_comparisons is enabled.