cpychecker: fix various format codes for Py_BuildValue
The format code handling for Py_BuildValue was missing support for the
following codes:

  'u' and 'u#'
  'f' and 'd'
  'D'
  'c'

In addition, the handling for 's#' and 'z#' had a bug in which it erroneously
expected an int* or Py_ssize_t*, rather than just a int or Py_ssize_t.

This commit fixes these issues, and (I hope) gives full coverage of all valid
format codes for Py_BuildValue in Python 2.