mass-rebuilds: work towards running cpychecker on large numbers of packages
Add new files:
  misc/fedora/bugreporting.py
  misc/fedora/bugreports.txt
to make it easier to file bugs into Fedora's bugzilla about issues found by
gcc-with-cpychecker.

Use this in mass-rebuild.py to keep track of which SRPMs we've already
reported vs those with only false positives etc, and to pop open "New Bug"
tabs in the browser with everything prepopulated when analyzing results
from the checker.

This reduces the amount of gruntwork needed when analyzing the results,
however it does still involve manually inspection before filing bugs.