misc/fedora: rework how we inject the plugin within the mass rebuild
Previously, I attempted to run the plugin by adding the necessary flags to
the __global_cflags macro.

However, this doesn't work for packages that don't use these build flags,
sometimes for the whole build, sometimes for just subsets of the build.

So instead, move /usr/bin/gcc to /usr/bin/the-real-gcc, and copy a harness
to /usr/bin/gcc; with similar changes for /usr/bin/g++

Hopefully this should make *everything* use the plugin during the mass rebuild