cpychecker: get the checker working on C++ code
Attempting to run the checker on C++ code would fail: it could not locate
any global declarations.

The logic to locate them in gccutils (searching for blocks in the
translation units) worked for C, but not for C++.

For C++, we add a new function:
   gcc.get_global_namespace()
to locate the gcc.NamespaceDecl for the "::" namespace, and add a lookup
method to gcc.NamespaceDecl.  We then use this for C++ to locate global
declarations, enabling the checker to run.

Also, introduce raise_cplusplus_only()