cpychecker: add various new fields when serializing values to JSON
Add "gcctype" and "value_comes_from" fields to AbstractValue.as_json, and
use a new per-class json_fields hook, so that subclasses can add extra fields.

Doing so requires RefcountValue to (at last) properly inherit from
AbstractValue, so we set the gcctype (hardcoded), and set the location
appropriately throughout.

In addition, add a RefcountValue.get_referrers_as_json() method, using this
to capture a JSON representation of the pointers pointing at the object.  To
do this requires adding a reference to the object to the RefcountValue.

Also, add a basic test of JSON serialization:
  tests/cpychecker/refcounts/json/basic
This only tests the serialization of states so far.